requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.42.0
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.25.0
pydantic>=2.0.0
pyyaml>=6.0
joblib>=1.3.0
python-multipart
requests
pytest
black
isort
flake8
scapy>=2.5.0
